+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2026-02-25
+
 ### Removed
 
 - Symbolics.jl extension (`ext/PlutoMathInputSymbolicsExt.jl`) and `:symbolics` output format